+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,13 +1,5 @@
 package com.example.example
 
-import android.os.Bundle
+import io.flutter.embedding.android.FlutterActivity
 
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
-
-class MainActivity: FlutterActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    GeneratedPluginRegistrant.registerWith(this)
-  }
-}
+class MainActivity: FlutterActivity() {}
