+++ lib/src/core/flexible_draggable_scrollable_sheet_scroll_controller.dart
@@ -1,0 +1,188 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:bottom_sheet/src/core/core.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+/// A [ScrollController] suitable for use in a [ScrollableWidgetBuilder] created
+/// by a [FlexibleDraggableScrollableSheet].
+///
+/// If a [FlexibleDraggableScrollableSheet] contains content that is exceeds
+/// the height
+/// of its container, this controller will allow the sheet to both be dragged to
+/// fill the container and then scroll the child content.
+///
+/// See also:
+///
+///  * [FlexibleDraggableScrollableSheetScrollPosition], which manages the
+/// positioning logic for
+///    this controller.
+///  * [PrimaryScrollController], which can be used to establish a
+///    [FlexibleDraggableScrollableSheetScrollController] as the primary
+/// controller for
+///    descendants.
+class FlexibleDraggableScrollableSheetScrollController
+    extends ScrollController {
+  final FlexibleDraggableSheetExtent extent;
+
+  FlexibleDraggableScrollableSheetScrollController({
+    required this.extent,
+    double initialScrollOffset = 0.0,
+    String? debugLabel,
+  }) : super(
+          debugLabel: debugLabel,
+          initialScrollOffset: initialScrollOffset,
+        );
+
+  @override
+  FlexibleDraggableScrollableSheetScrollPosition createScrollPosition(
+    ScrollPhysics physics,
+    ScrollContext context,
+    ScrollPosition? oldPosition,
+  ) {
+    return FlexibleDraggableScrollableSheetScrollPosition(
+      physics: physics,
+      context: context,
+      oldPosition: oldPosition,
+      extent: extent,
+    );
+  }
+
+  @override
+  void debugFillDescription(List<String> description) {
+    super.debugFillDescription(description);
+    description.add('extent: $extent');
+  }
+}
+
+/// A scroll position that manages scroll activities for
+/// [FlexibleDraggableScrollableSheetScrollController].
+///
+/// This class is a concrete subclass of [ScrollPosition] logic that handles a
+/// single [ScrollContext], such as a [Scrollable]. An instance of this class
+/// manages [ScrollActivity] instances, which changes the
+/// [FlexibleDraggableSheetExtent.currentExtent] or visible content offset in
+/// the [Scrollable]'s [Viewport]
+///
+/// See also:
+///
+///  * [FlexibleDraggableScrollableSheetScrollController], which uses this as
+/// its [ScrollPosition].
+class FlexibleDraggableScrollableSheetScrollPosition
+    extends ScrollPositionWithSingleContext {
+  final FlexibleDraggableSheetExtent extent;
+
+  bool get listShouldScroll => pixels > 0.0;
+
+  VoidCallback? _dragCancelCallback;
+
+  FlexibleDraggableScrollableSheetScrollPosition({
+    required ScrollPhysics physics,
+    required ScrollContext context,
+    required this.extent,
+    double initialPixels = 0.0,
+    bool keepScrollOffset = true,
+    ScrollPosition? oldPosition,
+    String? debugLabel,
+  }) : super(
+          physics: physics,
+          context: context,
+          initialPixels: initialPixels,
+          keepScrollOffset: keepScrollOffset,
+          oldPosition: oldPosition,
+          debugLabel: debugLabel,
+        );
+
+  @override
+  bool applyContentDimensions(double minScrollExtent, double maxScrollExtent) {
+    // We need to provide some extra extent if we haven't yet reached the max or
+    // min extents. Otherwise, a list with fewer children than the extent of
+    // the available space will get stuck.
+    return super.applyContentDimensions(
+      minScrollExtent - extent.additionalMinExtent,
+      maxScrollExtent + extent.additionalMaxExtent,
+    );
+  }
+
+  @override
+  void applyUserOffset(double delta) {
+    if (!listShouldScroll &&
+        (!(extent.isAtMin || extent.isAtMax) ||
+            (extent.isAtMin && delta < 0) ||
+            (extent.isAtMax && delta > 0))) {
+      extent.addPixelDelta(-delta, context.notificationContext);
+    } else {
+      super.applyUserOffset(delta);
+    }
+  }
+
+  @override
+  void goBallistic(double velocity) {
+    if (velocity == 0.0 ||
+        (velocity < 0.0 && listShouldScroll) ||
+        (velocity > 0.0 && extent.isAtMax)) {
+      super.goBallistic(velocity);
+      return;
+    }
+    //Scrollable expects that we will dispose of its current _dragCancelCallback
+    _dragCancelCallback?.call();
+    _dragCancelCallback = null;
+
+    // The iOS bouncing simulation just isn't right here - once we delegate
+    // the ballistic back to the ScrollView, it will use the right simulation.
+    final Simulation simulation = ClampingScrollSimulation(
+      position: extent.currentExtent,
+      velocity: velocity,
+      tolerance: physics.tolerance,
+    );
+
+    final ballisticController = AnimationController.unbounded(
+      debugLabel: '_FlexibleDraggableScrollableSheetScrollPosition',
+      vsync: context.vsync,
+    );
+    var lastDelta = 0.0;
+    void _tick() {
+      final delta = ballisticController.value - lastDelta;
+      lastDelta = ballisticController.value;
+      extent.addPixelDelta(delta, context.notificationContext);
+      if ((velocity > 0 && extent.isAtMax) ||
+          (velocity < 0 && extent.isAtMin)) {
+        // Make sure we pass along enough velocity to keep scrolling - otherwise
+        // we just "bounce" off the top making it look like the list doesn't
+        // have more to scroll.
+        // ignore: parameter_assignments
+        velocity = ballisticController.velocity +
+            (physics.tolerance.velocity * ballisticController.velocity.sign);
+        super.goBallistic(velocity);
+        ballisticController.stop();
+      } else if (ballisticController.isCompleted) {
+        super.goBallistic(0);
+      }
+    }
+
+    ballisticController
+      ..addListener(_tick)
+      ..animateWith(simulation).whenCompleteOrCancel(
+        ballisticController.dispose,
+      );
+  }
+
+  @override
+  Drag drag(DragStartDetails details, VoidCallback dragCancelCallback) {
+    // Save this so we can call it later if we have to [goBallistic] on our own.
+    _dragCancelCallback = dragCancelCallback;
+    return super.drag(details, dragCancelCallback);
+  }
+}
