+++ lib/src/core/reset_notifier.dart
@@ -1,0 +1,37 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:flutter/foundation.dart';
+
+/// A [ChangeNotifier] to use with InheritedResetNotifyer to notify
+/// descendants that they should reset to initial state.
+class ResetNotifier extends ChangeNotifier {
+  /// Whether someone called [sendReset] or not.
+  ///
+  /// This flag should be reset after checking it.
+  /// TODO: check should be private
+  bool wasCalled = false;
+
+  /// Fires a reset notification to descendants.
+  ///
+  /// Returns false if there are no listeners.
+  bool sendReset() {
+    if (!hasListeners) {
+      return false;
+    }
+    wasCalled = true;
+    notifyListeners();
+    return true;
+  }
+}
