+++ lib/src/flexible_bottom_sheet_header_delegate.dart
@@ -15,16 +15,8 @@
 import 'package:flutter/widgets.dart';
 
 class FlexibleBottomSheetHeaderDelegate extends SliverPersistentHeaderDelegate {
-  FlexibleBottomSheetHeaderDelegate({
-    required this.maxHeight,
-    required this.child,
-    this.minHeight = 0,
-  });
-
   final Widget child;
-
   final double minHeight;
-
   final double maxHeight;
 
   @override
@@ -32,6 +24,12 @@ class FlexibleBottomSheetHeaderDelegate extends SliverPersistentHeaderDelegate {
 
   @override
   double get maxExtent => maxHeight;
+
+  FlexibleBottomSheetHeaderDelegate({
+    required this.maxHeight,
+    required this.child,
+    this.minHeight = 0,
+  });
 
   @override
   bool shouldRebuild(SliverPersistentHeaderDelegate oldDelegate) => true;
